Implementacion de UI

+++ What Lies Beneath/Assets/Scripts/UISlider.cs
@@ -7,11 +7,11 @@ public class UISlider : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxValue(int maxValue) {
+    public void SetMaxValue(float maxValue) {
         slider.maxValue = maxValue;
     }
 
-    public void SetValue(int value) {
+    public void SetValue(float value) {
         slider.value = value;
     }
 }
